requirements.txt
aiogram>=3.4.0
pyyaml>=6.0
faster-whisper>=1.0.0
httpx>=0.27.0
